# PropertyValueConverter optimization: avoiding hashset lookup for primitive types